# 🚀 2.5x Speed Optimization: Hive Mind Coordination System