--- a/.github/workflows/ci-minimal.yml
+++ b/.github/workflows/ci-minimal.yml
@@ -1,4 +1,4 @@
-name: Emergency CI - Minimal Build
+name: Emergency CI - Minimal Build (Optimized)
 
 on:
   push:
@@ -8,18 +8,19 @@ on:
 
 env:
   NODE_VERSION: '20'
+  CACHE_VERSION: 'v2'
 
 jobs:
   emergency-build:
-    name: Emergency Build Check
+    name: Emergency Build Check (Fast)
     runs-on: ubuntu-latest
-    timeout-minutes: 10
+    timeout-minutes: 8
     
     steps:
       - name: Checkout code
         uses: actions/checkout@v4
 
-      - name: Setup Node.js
+      - name: Setup Node.js with enhanced caching
         uses: actions/setup-node@v4
         with:
           node-version: ${{ env.NODE_VERSION }}
@@ -28,30 +29,31 @@ jobs:
             backend/package-lock.json
             frontend/package-lock.json
 
-      - name: Install system dependencies
+      # Cache system dependencies for speed
+      - name: Cache system dependencies
+        uses: actions/cache@v4
+        with:
+          path: /var/cache/apt
+          key: ${{ env.CACHE_VERSION }}-${{ runner.os }}-system-deps
+
+      - name: Install system dependencies (cached)
         run: |
           sudo apt-get update
           sudo apt-get install -y python3 python3-dev make g++ sqlite3
 
-      - name: Install backend dependencies
-        working-directory: ./backend
-        run: npm ci --only=production
+      # Parallel dependency installation for speed
+      - name: Install dependencies in parallel
+        run: |
+          cd backend && npm ci --only=production --prefer-offline &
+          cd frontend && npm ci --only=production --prefer-offline &
+          wait
 
-      - name: Install frontend dependencies  
-        working-directory: ./frontend
-        run: npm ci --only=production
-
-      - name: Build backend (skip type errors for emergency)
-        working-directory: ./backend
-        run: npm run build || echo "Build failed - TypeScript errors found but continuing for emergency CI"
-
-      - name: Install missing frontend deps
-        working-directory: ./frontend
-        run: npm install lucide-react || echo "Install failed but continuing"
-
-      - name: Build frontend (skip errors for emergency)
-        working-directory: ./frontend
-        run: npm run build || echo "Build failed - dependency errors found but continuing for emergency CI"
+      # Parallel builds for maximum speed
+      - name: Build applications in parallel (emergency mode)
+        run: |
+          cd backend && (npm run build || echo "Backend build failed - continuing for emergency CI") &
+          cd frontend && npm install lucide-react && (npm run build || echo "Frontend build failed - continuing for emergency CI") &
+          wait
         env:
           NEXT_PUBLIC_API_URL: http://localhost:3001