--- a/backend/Dockerfile
+++ b/backend/Dockerfile
@@ -1,21 +1,108 @@
-FROM node:20-alpine
-
+# Multi-stage build for optimized Docker layers and caching
+# Stage 1: Dependency installation (cached layer)
+FROM node:20-alpine AS deps
 WORKDIR /app
 
-# Copy package files
+# Install system dependencies for better-sqlite3 and canvas
+RUN apk add --no-cache \
+    python3 \
+    make \
+    g++ \
+    cairo-dev \
+    jpeg-dev \
+    pango-dev \
+    musl-dev \
+    giflib-dev \
+    pixman-dev \
+    pangomm-dev \
+    libjpeg-turbo-dev \
+    freetype-dev
+
+# Copy package files for dependency installation
 COPY package*.json ./
 
-# Install dependencies
-RUN npm ci
+# Set environment variables to skip unnecessary downloads
+ENV PUPPETEER_SKIP_DOWNLOAD=true
+ENV PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD=true
+ENV CYPRESS_INSTALL_BINARY=0
+
+# Install only production dependencies, excluding dev packages that download browsers
+RUN npm ci --only=production --ignore-scripts && npm cache clean --force
+
+# Stage 2: Build stage (for TypeScript compilation)
+FROM node:20-alpine AS builder
+WORKDIR /app
+
+# Install build dependencies
+RUN apk add --no-cache \
+    python3 \
+    make \
+    g++ \
+    cairo-dev \
+    jpeg-dev \
+    pango-dev \
+    musl-dev \
+    giflib-dev \
+    pixman-dev \
+    pangomm-dev \
+    libjpeg-turbo-dev \
+    freetype-dev
+
+# Copy package files and install all dependencies (including dev)
+COPY package*.json ./
+
+# Set environment variables to skip unnecessary downloads
+ENV PUPPETEER_SKIP_DOWNLOAD=true
+ENV PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD=true
+ENV CYPRESS_INSTALL_BINARY=0
+
+# Install all dependencies for building, skipping post-install scripts for some packages
+RUN npm ci --ignore-scripts
 
 # Copy source code
 COPY . .
 
-# Create data directory
-RUN mkdir -p /app/data
+# Build the application
+RUN npm run build
+
+# Stage 3: Production runtime (minimal final image)
+FROM node:20-alpine AS runner
+WORKDIR /app
+
+# Install only runtime system dependencies
+RUN apk add --no-cache \
+    cairo \
+    jpeg \
+    pango \
+    musl \
+    giflib \
+    pixman \
+    libjpeg-turbo \
+    freetype
+
+# Create non-root user for security
+RUN addgroup --system --gid 1001 nodejs
+RUN adduser --system --uid 1001 nodeuser
+
+# Copy production dependencies from deps stage
+COPY --from=deps --chown=nodeuser:nodejs /app/node_modules ./node_modules
+
+# Copy built application from builder stage
+COPY --from=builder --chown=nodeuser:nodejs /app/dist ./dist
+COPY --from=builder --chown=nodeuser:nodejs /app/package*.json ./
+
+# Create data directory with proper permissions
+RUN mkdir -p /app/data && chown nodeuser:nodejs /app/data
+
+# Switch to non-root user
+USER nodeuser
 
 # Expose port
 EXPOSE 3001
 
-# Start development server
-CMD ["npm", "run", "dev"]
+# Health check for container orchestration
+HEALTHCHECK --interval=30s --timeout=10s --start-period=40s --retries=3 \
+  CMD node -e "require('http').get('http://localhost:3001/health', (res) => { process.exit(res.statusCode === 200 ? 0 : 1) }).on('error', () => process.exit(1))"
+
+# Start production server
+CMD ["npm", "start"]