--- a/frontend/.dockerignore
+++ b/frontend/.dockerignore
@@ -1,0 +1,109 @@
+# Node modules and package-lock (will be handled in Docker)
+node_modules
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+
+# Next.js build outputs
+.next
+out
+build
+dist
+
+# Development files
+.env.local
+.env.development
+.env.test
+.env*.local
+
+# IDE and editor files
+.vscode
+.idea
+*.swp
+*.swo
+*~
+
+# OS generated files
+.DS_Store
+.DS_Store?
+._*
+.Spotlight-V100
+.Trashes
+ehthumbs.db
+Thumbs.db
+
+# Git files
+.git
+.gitignore
+.gitattributes
+
+# Documentation and README files
+*.md
+docs/
+documentation/
+
+# Test files
+*.test.ts
+*.test.tsx
+*.test.js
+*.test.jsx
+*.spec.ts
+*.spec.tsx
+*.spec.js
+*.spec.jsx
+__tests__/
+coverage/
+.nyc_output
+playwright-report/
+test-results/
+
+# Logs
+logs
+*.log
+
+# Runtime data
+pids
+*.pid
+*.seed
+*.pid.lock
+
+# Coverage directory used by tools like istanbul
+coverage/
+.nyc_output
+
+# TypeScript cache
+*.tsbuildinfo
+
+# Optional npm cache directory
+.npm
+
+# Optional REPL history
+.node_repl_history
+
+# Output of 'npm pack'
+*.tgz
+
+# Yarn Integrity file
+.yarn-integrity
+
+# dotenv environment variables file
+.env
+
+# Dependency directories
+jspm_packages/
+
+# Optional eslint cache
+.eslintcache
+
+# Optional cache data for Prettier
+.prettiercache
+
+# Storybook build outputs
+storybook-static
+
+# Claude Flow cache and temporary files
+.swarm/
+.claude/
+
+# Playwright
+playwright/.cache/