--- a/frontend/Dockerfile
+++ b/frontend/Dockerfile
@@ -1,18 +1,74 @@
-FROM node:20-alpine
-
+# Multi-stage build for optimized Next.js application
+# Stage 1: Dependency installation (cached layer)
+FROM node:20-alpine AS deps
 WORKDIR /app
 
-# Copy package files
+# Copy package files for dependency installation
 COPY package*.json ./
 
-# Install dependencies
-RUN npm ci
+# Set environment variables to skip unnecessary downloads
+ENV PUPPETEER_SKIP_DOWNLOAD=true
+ENV PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD=true
+ENV CYPRESS_INSTALL_BINARY=0
+
+# Install only production dependencies
+RUN npm ci --only=production --ignore-scripts && npm cache clean --force
+
+# Stage 2: Build stage (for Next.js compilation)
+FROM node:20-alpine AS builder
+WORKDIR /app
+
+# Copy package files and install all dependencies (including dev)
+COPY package*.json ./
+
+# Set environment variables to skip unnecessary downloads
+ENV PUPPETEER_SKIP_DOWNLOAD=true
+ENV PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD=true
+ENV CYPRESS_INSTALL_BINARY=0
+
+# Install all dependencies for building
+RUN npm ci --ignore-scripts
 
 # Copy source code
 COPY . .
 
+# Set build environment variables
+ENV NODE_ENV=production
+ENV NEXT_TELEMETRY_DISABLED=1
+
+# Build the Next.js application
+RUN npm run build
+
+# Stage 3: Production runtime (minimal final image)
+FROM node:20-alpine AS runner
+WORKDIR /app
+
+# Create non-root user for security
+RUN addgroup --system --gid 1001 nodejs
+RUN adduser --system --uid 1001 nextjs
+
+# Set production environment
+ENV NODE_ENV=production
+ENV NEXT_TELEMETRY_DISABLED=1
+
+# Copy production dependencies from deps stage
+COPY --from=deps --chown=nextjs:nodejs /app/node_modules ./node_modules
+
+# Copy built application from builder stage
+COPY --from=builder --chown=nextjs:nodejs /app/.next/standalone ./
+COPY --from=builder --chown=nextjs:nodejs /app/.next/static ./.next/static
+COPY --from=builder --chown=nextjs:nodejs /app/public ./public
+COPY --from=builder --chown=nextjs:nodejs /app/package*.json ./
+
+# Switch to non-root user
+USER nextjs
+
 # Expose port
 EXPOSE 3000
 
-# Start development server
-CMD ["npm", "run", "dev"]
+# Health check for container orchestration
+HEALTHCHECK --interval=30s --timeout=10s --start-period=40s --retries=3 \
+  CMD node -e "require('http').get('http://localhost:3000', (res) => { process.exit(res.statusCode === 200 ? 0 : 1) }).on('error', () => process.exit(1))"
+
+# Start production server
+CMD ["npm", "start"]