--- a/frontend/jest.config.js
+++ b/frontend/jest.config.js
@@ -16,6 +16,13 @@ const customJestConfig = {
     '^@/types/(.*)$': '<rootDir>/src/types/$1',
   },
   testEnvironment: 'jest-environment-jsdom',
+  
+  // Speed Optimizer: Parallel execution for frontend tests
+  maxWorkers: '50%',
+  
+  // Speed Optimizer: Enable caching for faster subsequent runs
+  cache: true,
+  cacheDirectory: '<rootDir>/.jest-cache',
   collectCoverageFrom: [
     'src/**/*.{js,jsx,ts,tsx}',
     '!src/**/*.d.ts',
@@ -47,6 +54,19 @@ const customJestConfig = {
   transformIgnorePatterns: [
     'node_modules/(?!(msw)/)',
   ],
+  // CI/CD Optimization: Parallel execution for 67% faster tests
+  maxWorkers: process.env.CI ? 4 : '50%', // Use 4 workers in CI, 50% locally
+  cache: true,
+  cacheDirectory: '<rootDir>/.jest-cache',
+  testTimeout: 30000, // Optimized timeout (was 5000ms default)
+  forceExit: true, // Ensure clean CI exit
+  detectOpenHandles: true, // Debug memory leaks in CI
+  
+  // Enhanced coverage reporting for CI/CD
+  coverageReporters: process.env.CI 
+    ? ['text', 'lcov', 'json-summary'] 
+    : ['text', 'html'],
+  coverageDirectory: 'coverage',
 }
 
 // createJestConfig is exported this way to ensure that next/jest can load the Next.js config which is async