--- a/frontend/next.config.js
+++ b/frontend/next.config.js
@@ -2,9 +2,18 @@
 const nextConfig = {
   reactStrictMode: true,
   swcMinify: true,
+  // Enable standalone output for optimized Docker builds
+  output: 'standalone',
+  // Compress output for smaller image size
+  compress: true,
+  // Optimize images for better performance
+  images: {
+    unoptimized: false,
+  },
+  // Enable experimental features for better performance
   experimental: {
-    // appDir is no longer experimental in Next.js 13.4+
-    // Remove this line or use app directory features directly
+    // Enable optimized package imports
+    optimizePackageImports: ['react', 'react-dom'],
   },
   env: {
     NEXT_PUBLIC_API_URL: process.env.NEXT_PUBLIC_API_URL || 'http://localhost:3001',