--- a/frontend/package.json
+++ b/frontend/package.json
@@ -12,7 +12,7 @@
     "test:unit": "jest --testPathPattern=__tests__ --passWithNoTests",
     "test:watch": "jest --watch",
     "test:coverage": "jest --coverage --watchAll=false",
-    "test:ci": "jest --coverage --watchAll=false --testTimeout=30000",
+    "test:ci": "jest --coverage --watchAll=false --testTimeout=15000",
     "test:e2e": "playwright test",
     "test:e2e:ui": "playwright test --ui", 
     "test:e2e:headed": "playwright test --headed",