… .github/workflows/ci-optimized.yml
… backend/package.json
… docker-compose.yml